+++ lib/waterdrop/patches/rdkafka/metadata.rb
@@ -7,6 +7,14 @@ module WaterDrop
     module Rdkafka
       # Rdkafka::Metadata patches
       module Metadata
+        # Errors upon which we retry the metadata fetch
+        RETRIED_ERRORS = %i[
+          timed_out
+          leader_not_available
+        ].freeze
+
+        private_constant :RETRIED_ERRORS
+
         # We overwrite this method because there were reports of metadata operation timing out
         # when Kafka was under stress. While the messages dispatch will be retried, metadata
         # fetch happens prior to that, effectively crashing the process. Metadata fetch was not
@@ -19,7 +27,7 @@ module WaterDrop
 
           super(*args)
         rescue ::Rdkafka::RdkafkaError => e
-          raise unless e.code == :timed_out
+          raise unless RETRIED_ERRORS.include?(e.code)
           raise if attempt > 10
 
           backoff_factor = 2**attempt